pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "my stepping stone to the web developement world"
authors = ["midnight <ivanator666@gmail.com>"]
license = "GPL-3"
readme = "README.md"
packages = [
    {include = "page_analyzer"}
]
classifiers = [
    "Environment :: Console",
    "Programming Language :: Python :: 3",
    "Topic :: Utilities"
    ]
repository = "https://github.com/Midnight95/python-project-83.git"


[tool.poetry.dependencies]
python = ">=3.10"
gunicorn = "^20.1.0"
flask = "^2.3.2"
psycopg2-binary = "^2.9.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
flake8 = "^6.0.0"
coverage = "^7.2.5"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
